+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Automatic refresh of data now executed one minute after cron event has executed, to let time to do the actions
+
 ### 1.4.4: May 27th, 2021
 * Dialog Insight: Make birthdate value optional when sending to web service
 * Markup "modifyUrl" now goes to cancel page, not contact page