+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * Automatic refresh of data now executed one minute after cron event has executed, to let time to do the actions
 * Mini dashboard: Remove javascript, fix margins, add link to main dashboard
 * Fix missing aquos_generate_signature function
+* Prevent non-admins from seeing  dashboard calculations
 
 ### 1.4.4: May 27th, 2021
 * Dialog Insight: Make birthdate value optional when sending to web service