+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+### 1.4.5: May 31th, 2021
 * Automatic refresh of data now executed one minute after cron event has executed, to let time to do the actions
 * Mini dashboard: Remove javascript, fix margins, add link to main dashboard
 * Fix missing aquos_generate_signature function